mirai.promises: Make 'Mirai' 'Promises'

Allows 'mirai' objects encapsulating asynchronous computations, from the mirai package by Gao (2023), to be used interchangeably with 'promise' objects from the promises package by Cheng (2021). This facilitates their use with packages plumber by Schloerke and Allen (2022) and shiny by Cheng, Allaire, Sievert, Schloerke, Xie, Allen, McPherson, Dipert and Borges (2022).

Notes

This package provides the methods as.promise.mirai and as.promise.recvAio for the S3 generic as.promise exported by the 'promises' package.

An auxiliary function polling provides the additional option to tune the frequency at which 'mirai' are checked for resolution.

Package authors wishing to use the S3 methods may simply import the function polling to make them available.

Links

mirai website: https://shikokuchuo.net/mirai/

nanonext website: https://shikokuchuo.net/nanonext/

NNG website: https://nng.nanomsg.org/

Author

Charlie Gao charlie.gao@shikokuchuo.net (ORCID)